==> 1_eda_prepare_dataset.ipynb <==
import pandas as pd

# 1-  Cargar archivo Excel
xlsx_path = "customer-details.xlsx"
excel_file = pd.ExcelFile(xlsx_path)

# 🔍 Ver nombres de hojas
print("📄 Hojas disponibles:", excel_file.sheet_names)


# Mapeo hoja -> año
mapa_hojas = {
    '2012': '2012',
    '2013': '2013',
    '2014': '2014'
}

# 🔀 Unificar todas las hojas en un solo DataFrame con columna 'año'
df_xlsx_all = pd.concat(
    [excel_file.parse(sheet).assign(año=year) for sheet, year in mapa_hojas.items()],
    ignore_index=True
)



# ✅ Renombrar columna 'ID' → 'uuid' para preparar el merge
df_xlsx_all = df_xlsx_all.rename(columns={'ID': 'uuid'})
# 🧪 Comprobar resultado
print("📎 Dataset combinado desde el Excel:")
print(f"🧮 Registros: {df_xlsx_all.shape[0]}")
print(f"📋 Columnas: {df_xlsx_all.columns.tolist()}")
print("🧪 ¿Está 'uuid' en df_xlsx_all?", 'uuid' in df_xlsx_all.columns)
print("📋 Primeras filas:", df_xlsx_all[['uuid']].head())

# 2 — Cargar los datos CSV
with open("bank-additional.csv", "r", encoding="utf-8") as f:
    cabecera_raw = f.readline().strip()
if cabecera_raw.startswith(","):
    cabecera_raw = cabecera_raw[1:]
    columnas = [col.strip() for col in cabecera_raw.split(",")]
try:
    df_bank  = pd.read_csv(
        "bank-additional.csv",
        sep=",",
        skiprows=1,
        names=columnas,
        encoding="utf-8",
        quotechar='"',
        decimal=","
    )
    print("✅ CSV cargado con éxito")
except Exception as e:
    print("❌ Error al cargar el CSV:", e)
print(f"🔁 Registros completamente duplicados (todas las columnas): {duplicados_completos}")

# Renombrar columna id_ → uuid si aún no está hecho
df_bank = df_bank.rename(columns={'id_': 'uuid'})

print("📎 Dataset CSV:")
print(f"🧮 Registros: {df_bank.shape[0]}")
print(f"📋 Columnas: {df_bank.columns.tolist()}")


# 3 - Mergeamos

# 🔗 Realizamos el merge usando 'uuid' como clave
df_merged = pd.merge(df_bank, df_xlsx_all, on='uuid', how='outer')

# 🧪 Verificamos el resultado del merge
print("✅ Merge completado con éxito")
print(f"🔢 Registros en df_bank: {df_bank.shape[0]}")
print(f"🔢 Registros en df_xlsx_all: {df_xlsx_all.shape[0]}")
print(f"🔗 Registros después del merge: {df_merged.shape[0]}")
print(f"📋 Columnas del nuevo dataset: {df_merged.columns.tolist()}")

# 👁️ Mostramos algunas filas para verificar visualmente
display(df_merged.sample(5))

# 4 - Exportar DataFrame fusionado a archivo Excel
output_path = "data_set_complete_raw.xlsx"

# 💾 Guardar sin índice, para mantenerlo limpio
df_merged.to_excel(output_path, index=False)

# ✅ Confirmación
print(f"✅ Archivo guardado con éxito en: {output_path}")
