==> 4_eda_proyecto_KPI_2.ipynb <==
KPI 2: Arquetipo de cliente que contrata productos bancarios.

⸻

🎯 Objetivo

Identificar el perfil demográfico y financiero más representativo de los clientes que sí han contratado un producto (y == True).



import pandas as pd
import importlib
import data_cleaning_utils
importlib.reload(data_cleaning_utils)

# 1 - Cargar el DataFrame
df_bank = pd.read_excel("data_set_complete.xlsx")
df_bank.info()

# Paso 1 - Crear variable de trabajo consolidada
df_arquetipo = df_bank[df_bank['y'] == True].copy()

# Paso 2 - Seleccionar columnas relevantes (puedes añadir más si las necesitas luego)
columnas_kpi2 = [
    'age', 'job', 'marital', 'education',
    'housing', 'loan', 'Income', 'Kidhome', 'Teenhome', 'contact'
]
df_arquetipo = df_arquetipo[columnas_kpi2]

# Paso 3 - Añadir columnas derivadas
df_arquetipo['children'] = df_arquetipo['Kidhome'] + df_arquetipo['Teenhome']
df_arquetipo['has_children'] = df_arquetipo['children'].apply(lambda x: 1 if x > 0 else 0)

# Llamada a la función con la variable 'age' e 'Income
data_cleaning_utils.resumen_variable_numerica(df_arquetipo, 'age')
data_cleaning_utils.resumen_variable_numerica(df_arquetipo, 'Income')

data_cleaning_utils.resumen_variable_categorica(df_arquetipo, 'job')
data_cleaning_utils.resumen_variable_categorica(df_arquetipo, 'marital')
data_cleaning_utils.resumen_variable_categorica(df_arquetipo, 'education')
data_cleaning_utils.resumen_variable_categorica(df_arquetipo, 'housing')
data_cleaning_utils.resumen_variable_categorica(df_arquetipo, 'loan')
data_cleaning_utils.resumen_variable_categorica(df_arquetipo, 'contact')

🧩 Resumen del arquetipo completo (KPI 2)

El cliente que contrata productos bancarios tiene una media de 39 años, ingresos cercanos a 92.000 € anuales, está casado, tiene estudios universitarios, trabaja en administración, tiene hipoteca, no tiene préstamo personal, tiene hijos y fue contactado por teléfono móvil.
